cmd/sanguisuga: fix issue where torrents would start stoned
For some reason when I started torrents, they would just not work. This
fixes the torrent downloader by having it set the b64 torrent meta-info
instead of passing the URL to the Transmission server. This seems to
make torrents actually start downloading.

It also makes this use one of Dominikh's libraries instead of the one
that the Transmission RPC spec suggested.

Signed-off-by: Xe Iaso <me@xeiaso.net>