fix(pvfm/bot): cache seen messages for up to 10 minutes
There is an issue in the PonyvilleFM Discord where the bot seems to
sometimes process commands twice in a row. I think this is causing some
spooky nondeterminism due to the fact that I'm assuming that Discord
delivers messages to me exactly once.

This is kind of a hack, but this may just work. I don't know.

Signed-off-by: Xe Iaso <me@xeiaso.net>