feat(aura): build with docker buildx bake
Listen to me for my tale of woe:

I was an idiot. When I moved this to the non-Earthly build process, I
slapped it into a container image without anything but the aura binary.
I had forgotten that aura explicitly depends on the streamripper binary
being installed. This is fundamentally how it records streams. Without
this binary, the bot does not work.

In testing, I found out that the version of streamripper in homebrew is
catastrophically broken. It will attempt to run, but all runs result in
failure. The version of streamripper in Debian is the only one that
seems to work.

I will need to rotate aura's discord key at some point, but for now I
can live with this as-is.

Release-Status: cut
Signed-off-by: Xe Iaso <me@xeiaso.net>